fix(hypervisor,compute): Display impls, serde derives, adapter stubs
* Add `impl Display for RegistryType` and `PhaseEra` in registry/mod.rs
  so `info!("{}", registry.registry_type())` calls in
  hybrid_master_registry.rs compile.
* registry_adapters.rs: add `use serde::de::DeserializeOwned`, add
  `Send + Sync` to the UnifiedRegistryAdapter<T> trait bounds, fix
  `entry.meta.health` by mapping unified_registry::EntryHealth to
  registry::EntryHealth, fix `e.id == *id` borrow issue, and stub
  out query_entity for adapters whose underlying types have
  private or non-existent fields (FederationModelRegistry, ModelRegistry,
  ComponentRegistry, SpecialistRegistry, ChromosomeRegistry,
  DistributedSpecialistRegistry). HoxCapabilityRegistryAdapter is
  reworked to use the public HoxRegistry API.
* dashboard.rs: clone the AlertThreshold before passing it to
  check_threshold to release the immutable borrow on
  self.config.alert_thresholds.
* batch_processor.rs: add `#[serde(skip, default = "Instant::now")]`
  to the Instant field in BatchedTask so the struct derives Serialize.
* components/compute/src/thermodynamics.rs: add `Serialize, Deserialize`
  to SystemPhase so it can be (de)serialized inside UnifiedSystemState.
* unified_learning.rs: add `Serialize, Deserialize` to
  UnifiedSystemState, fix f32/f64 mismatches in
  update_specialist_metabolism and ModelParameters.

+++ core/hypervisor/src/unified_learning.rs
@@ -22,7 +22,7 @@ use serde::{Deserialize, Serialize};
 
 /// Unified system state.
 /// Single state vector that all components operate on.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct UnifiedSystemState {
     // Thermodynamics
     pub free_energy: f64,
@@ -373,8 +373,8 @@ impl UnifiedLearningLoop {
     fn update_specialist_metabolism(&mut self, specialist_idx: usize, dopamine_reward: f32, confidence: f32) {
         // Increase metabolism (activation) for well-performing specialists
         if dopamine_reward > 0.2 {
-            self.biology.expression_rate = (self.biology.expression_rate + (dopamine_reward as f64 * 0.1))
-                .min(2.0);
+            self.biology.expression_rate = (self.biology.expression_rate + (dopamine_reward as f64 * 0.1) as f32)
+                .min(2.0_f32);
         }
         
         // Decrease metabolism for poorly-performing specialists
@@ -428,7 +428,7 @@ impl UnifiedLearningLoop {
             tensor_router_weights: self.tensor_router.weights.weights.clone(),
             learning_rate: self.config.learning_rate,
             predictive_coding_layers: self.predictive_coding.layers.len(),
-            biology_expression_rate: self.biology.expression_rate,
+            biology_expression_rate: self.biology.expression_rate as f64,
             system_state: self.system_state.clone(),
         }
     }
